Multiple lasers and damage
Created code for multiple lasers that do damage to enemies (damage might be a touch bugged)

+++ 2DSpaceShooter/Assets/_Scripts/Interfaces/IDamageable.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void TakeDamage(int amount);
+    void TakeDamage(float amount);
 }
